Pass through events on Android
