+++ MediaManager/Plugin.MediaManager.Android/Plugin.MediaManager.Android.csproj
@@ -69,7 +69,6 @@
     <Compile Include="MusicBroadcastReceiver.cs" />
     <Compile Include="RemoteControlBroadcastReceiver.cs" />
     <Compile Include="AudioPlayerImplementation.cs" />
-    <Compile Include="MediaPlayerServiceConnection.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Resources\AboutResources.txt" />
